Organize UnitInvokeTests
