+++ src/core-unit/Unit.Tests/UnitInvokeTests/Invoke.01.cs
@@ -9,7 +9,7 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Core.Tests
 {
-    partial class UnitTests
+    partial class UnitInvokeTests
     {
         [Test]
         public void Invoke_01_ActionIsNull_ExpectArgumentNullException()
