+++ src/core-unit/Unit.Tests/UnitInvokeTests/Invoke.04.cs
@@ -9,7 +9,7 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Core.Tests
 {
-    partial class UnitTests
+    partial class UnitInvokeTests
     {
         [Test]
         public void Invoke_04_ActionIsNull_ExpectArgumentNullException()
