+++ src/core-unit/Unit.Tests/UnitInvokeTests/Invoke.06.cs
@@ -9,38 +9,40 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Core.Tests
 {
-    partial class UnitTests
+    partial class UnitInvokeTests
     {
         [Test]
-        public void Invoke_05_ActionIsNull_ExpectArgumentNullException()
+        public void Invoke_06_ActionIsNull_ExpectArgumentNullException()
         {
-            Action<StructType, RefType, string, int, object> action = null!;
+            Action<StructType, RefType, string, int, object, DateTime> action = null!;
 
             var arg1 = SomeTextStructType;
             var arg2 = PlusFifteenIdRefType;
             var arg3 = TabString;
             var arg4 = MinusFortyFive;
             var arg5 = new { Value = PlusTwoHundredPointFive };
+            var arg6 = Year2015March11H01Min15;
 
-            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5));
+            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5, arg6));
             Assert.AreEqual("action", ex!.ParamName);
         }
 
         [Test]
-        public void Invoke_05_ExpectCallActionOnce()
+        public void Invoke_06_ExpectCallActionOnce()
         {
-            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?>();
+            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?, DateTime>();
 
             var arg1 = SomeTextStructType;
             var arg2 = (RefType?)null;
             var arg3 = TabString;
             var arg4 = MinusFortyFive;
             var arg5 = new { Value = PlusTwoHundredPointFive };
+            var arg6 = Year2015March11H01Min15;
 
-            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5);
+            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5, arg6);
 
             Assert.AreEqual(Unit.Value, actual);
-            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5), Times.Once);
+            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5, arg6), Times.Once);
         }
     }
 }
