+++ src/core-unit/Unit.Tests/UnitInvokeTests/Invoke.07.cs
@@ -9,12 +9,12 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Core.Tests
 {
-    partial class UnitTests
+    partial class UnitInvokeTests
     {
         [Test]
-        public void Invoke_09_ActionIsNull_ExpectArgumentNullException()
+        public void Invoke_07_ActionIsNull_ExpectArgumentNullException()
         {
-            Action<StructType, RefType, string, int, object, DateTime, StructType?, decimal, RefType> action = null!;
+            Action<StructType, RefType, string, int, object, DateTime, StructType?> action = null!;
 
             var arg1 = SomeTextStructType;
             var arg2 = PlusFifteenIdRefType;
@@ -23,17 +23,15 @@ namespace PrimeFuncPack.Core.Tests
             var arg5 = new { Value = PlusTwoHundredPointFive };
             var arg6 = Year2015March11H01Min15;
             var arg7 = NullTextStructType;
-            var arg8 = MinusSeventyFivePointSeven;
-            var arg9 = ZeroIdRefType;
 
-            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9));
+            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5, arg6, arg7));
             Assert.AreEqual("action", ex!.ParamName);
         }
 
         [Test]
-        public void Invoke_09_ExpectCallActionOnce()
+        public void Invoke_07_ExpectCallActionOnce()
         {
-            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?, DateTime, StructType?, decimal?, RefType>();
+            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?, DateTime, StructType?>();
 
             var arg1 = SomeTextStructType;
             var arg2 = (RefType?)null;
@@ -42,13 +40,11 @@ namespace PrimeFuncPack.Core.Tests
             var arg5 = new { Value = PlusTwoHundredPointFive };
             var arg6 = Year2015March11H01Min15;
             var arg7 = (StructType?)null;
-            var arg8 = (decimal?)MinusSeventyFivePointSeven;
-            var arg9 = ZeroIdRefType;
 
-            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9);
+            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5, arg6, arg7);
 
             Assert.AreEqual(Unit.Value, actual);
-            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9), Times.Once);
+            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7), Times.Once);
         }
     }
 }
