+++ src/core-unit/Unit.Tests/UnitInvokeTests/Invoke.09.cs
@@ -9,12 +9,12 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Core.Tests
 {
-    partial class UnitTests
+    partial class UnitInvokeTests
     {
         [Test]
-        public void Invoke_08_ActionIsNull_ExpectArgumentNullException()
+        public void Invoke_09_ActionIsNull_ExpectArgumentNullException()
         {
-            Action<StructType, RefType, string, int, object, DateTime, StructType?, decimal> action = null!;
+            Action<StructType, RefType, string, int, object, DateTime, StructType?, decimal, RefType> action = null!;
 
             var arg1 = SomeTextStructType;
             var arg2 = PlusFifteenIdRefType;
@@ -24,15 +24,16 @@ namespace PrimeFuncPack.Core.Tests
             var arg6 = Year2015March11H01Min15;
             var arg7 = NullTextStructType;
             var arg8 = MinusSeventyFivePointSeven;
+            var arg9 = ZeroIdRefType;
 
-            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8));
+            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9));
             Assert.AreEqual("action", ex!.ParamName);
         }
 
         [Test]
-        public void Invoke_08_ExpectCallActionOnce()
+        public void Invoke_09_ExpectCallActionOnce()
         {
-            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?, DateTime, StructType?, decimal?>();
+            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?, DateTime, StructType?, decimal?, RefType>();
 
             var arg1 = SomeTextStructType;
             var arg2 = (RefType?)null;
@@ -42,11 +43,12 @@ namespace PrimeFuncPack.Core.Tests
             var arg6 = Year2015March11H01Min15;
             var arg7 = (StructType?)null;
             var arg8 = (decimal?)MinusSeventyFivePointSeven;
+            var arg9 = ZeroIdRefType;
 
-            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8);
+            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9);
 
             Assert.AreEqual(Unit.Value, actual);
-            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8), Times.Once);
+            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9), Times.Once);
         }
     }
 }
