+++ src/core-unit/Unit.Tests/UnitInvokeTests/Invoke.10.cs
@@ -9,12 +9,12 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Core.Tests
 {
-    partial class UnitTests
+    partial class UnitInvokeTests
     {
         [Test]
-        public void Invoke_13_ActionIsNull_ExpectArgumentNullException()
+        public void Invoke_10_ActionIsNull_ExpectArgumentNullException()
         {
-            Action<StructType, RefType, string, int, object, DateTime, StructType?, decimal, RefType, object, StructType, string, double> action = null!;
+            Action<StructType, RefType, string, int, object, DateTime, StructType?, decimal, RefType, object> action = null!;
 
             var arg1 = SomeTextStructType;
             var arg2 = PlusFifteenIdRefType;
@@ -26,18 +26,15 @@ namespace PrimeFuncPack.Core.Tests
             var arg8 = MinusSeventyFivePointSeven;
             var arg9 = ZeroIdRefType;
             var arg10 = new object();
-            var arg11 = CustomStringStructType;
-            var arg12 = CustomText;
-            var arg13 = PlusFortyOnePointSeventyFive;
 
-            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11, arg12, arg13));
+            var ex = Assert.Throws<ArgumentNullException>(() => _ = Unit.Invoke(action, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10));
             Assert.AreEqual("action", ex!.ParamName);
         }
 
         [Test]
-        public void Invoke_13_ExpectCallActionOnce()
+        public void Invoke_10_ExpectCallActionOnce()
         {
-            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?, DateTime, StructType?, decimal?, RefType, object, StructType, string, double>();
+            var mockAction = MockActionFactory.CreateMockAction<StructType, RefType?, string, int, object?, DateTime, StructType?, decimal?, RefType, object>();
 
             var arg1 = SomeTextStructType;
             var arg2 = (RefType?)null;
@@ -49,14 +46,11 @@ namespace PrimeFuncPack.Core.Tests
             var arg8 = (decimal?)MinusSeventyFivePointSeven;
             var arg9 = ZeroIdRefType;
             var arg10 = new object();
-            var arg11 = CustomStringStructType;
-            var arg12 = CustomText;
-            var arg13 = PlusFortyOnePointSeventyFive;
 
-            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11, arg12, arg13);
+            var actual = Unit.Invoke(mockAction.Object.Invoke, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10);
 
             Assert.AreEqual(Unit.Value, actual);
-            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11, arg12, arg13), Times.Once);
+            mockAction.Verify(a => a.Invoke(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10), Times.Once);
         }
     }
 }
